requirements.txt
# Core dependencies
pydantic>=2.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0

# OCR
easyocr>=1.7.0
paddleocr>=2.7.0
paddlepaddle>=2.5.0
pytesseract>=0.3.10
Pillow>=10.0.0
pdf2image>=1.16.3
pypdf>=3.17.0

# LLM (Ollama via HTTP)
requests>=2.31.0
langchain>=0.1.0
langchain-classic>=0.1.0
langchain-community>=0.0.10
langchain-ollama>=0.0.1

# YOLO Food Recognition
ultralytics>=8.0.0

# CLIP Food Classification
transformers>=4.40.0
torch>=2.0.0
pandas>=2.0.0
sentencepiece>=0.1.99
protobuf>=4.21.0

# OpenCLIP for food classification
open_clip_torch>=2.20.0

# Vector Database for semantic search
chromadb>=0.4.0

